refactor(chuang-api): 优化接口文档和代码结构
-为控制器类和方法添加详细的 Swagger 注释
- 优化代码格式和导入顺序
- 为请求参数添加 Parameter 注解和描述- 修正部分方法的返回类型和参数类型

+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/service/SysDeptService.java
@@ -46,7 +46,7 @@ public interface SysDeptService extends IService<SysDept> {
      * @param id 部门ID
      * @return 部门信息
      */
-    SysDept getDeptById(Integer id);
+    SysDept getDeptById(Long id);
 
 
     /**
@@ -62,7 +62,7 @@ public interface SysDeptService extends IService<SysDept> {
      *
      * @param id 部门ID
      */
-    boolean deptHasSubordinates(Integer id);
+    boolean deptHasSubordinates(Long id);
 
     /**
      * 删除部门，支持批量删除
@@ -70,7 +70,7 @@ public interface SysDeptService extends IService<SysDept> {
      * @param ids 部门ID集合
      * @return 操作结果
      */
-    boolean removeDeptById(List<Integer> ids);
+    boolean removeDeptById(List<Long> ids);
 
     /**
      * 构建部门树
